Moved to ci.aquarium instead of using com for java (#7)
Changed the package to ci.aquarium instead of using com prefix. Also fixed checking for text files on linux.

I think it's a good timing because after release we will have issues with compatibility, so it's better to change right away, since Adobe has no intention to continue support of the system.